fixes #1056
Menu's and Flyout's set max-width to 456px, but this is not adaptive enough for phones that apply a zoom factor making some phones 320px wide in virtual pixels.
The fix is to use a media query that changes the max width of the menu and flyout to 100vw when the width of the media is less than 456px.